compose/chat-api/requirements_enterprise.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
redis==5.0.7
httpx==0.27.0
boto3==1.34.148
python-dotenv==1.0.1

# Enterprise Features
PyJWT==2.8.0
prometheus-client==0.19.0
python-multipart==0.0.6
pydantic==2.5.0

# Performance & Monitoring
aioredis==2.0.1
aiofiles==23.2.1
structlog==23.2.0

# Queue System
celery==5.3.4
kombu==5.3.4

# Security
cryptography==41.0.8
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Additional AI Models
openai==1.3.7
anthropic==0.7.8
